SC-66/feat: Added function to get contacts as HashMap
